Change the way we prevent iOS from adding its special dark styling to form elements
iOS makes form elements lighter and more rounded if they have a black background.
We don't want that on our buttons because they become hard to read, so we apply
-webkit-appearance: none to certain elements. We have to be specific to prevent
checkboxes, radios, and other potentially useful things from vanishing.

TODO: Not in this pull request, but the default skin needs the same declaration
on the button for closing the admin banner.