AO3-5790 Fix work and bookmark searches for certain languages
With thanks to Elz for both figuring out what the issue was, and for the solution.

When we changed the language search to use the language code instead of the numeric ID,
we didn't set the mapping for that field to `keyword`.

This has worked ok for languages with simple codes having the same capitalisation throughout, but has broken down for languages with mixed capitalisation in their codes, such as português brasileiro (`ptBR`) and português europeu (`ptPT`).

The most elegant solution would be to set the mapping and reindex, but we don't want to do another reindex so soon after we just did a big one, so what we are doing is change the filter.