AO3-2240 Avoid stale data issues
Before sending gift notifications, we want to be sure the work isn't unrevealed
because it's in another collection. Previously, I tried to check the item's
unrevealed status in the mailer. However, if the item's revealed status wasn't
updated immediately, the email wouldn't send.

Now I'm more explicit about what I want to check: in the callback that handles
notifications, the code checks if the item is in any unrevealed collections
other than the one where it is currently being revealed.