AO3-5813 Use Faker for work titles in user_mailer_spec
The work titles in the test were previously set up in such a way
that even if the work with the first title wasn't being listed,
the tests would still pass (because the second title contained
the full text of the first). Further, the text of the first title
was also in the email body.