AO3-5813 Remove HTML for mailers from locale file (#3709)
* AO3-5813 Remove html sent_by key from locale file

This key is only used by a mailer that is no longer used. Ergo
the mailer and its code should go bye-bye too. Admin notification
feature was removed in AO3-5043.

* AO3-5813 Remove HTML from footer mailer text in locale

* AO3-5813 Don't include italics in kudo mailer

We reuse the italic bold link style a lot for titles in emails.
Let's make it a helper.

* AO3-5813 Remove HTML for invite_increase_notification from mailer locale

* AO3-5813 Improve handling of different numbers in invite_increase_notification

Also capitalize Archive because it made Priscilla and me twitch

* AO3-5813 Update specs to be consistent with other reworked mailer specs

* AO3-5813 Spacing fixes in spec files

* AO3-5813 Prefer double quotes

* AO3-5813 Update the text we check for in the invite_increase_notification tests

We don't want any false successes, so let's check for 'invitations, which'
when there is >1 invitation, and 'invitation, which' when there is only 1.

* AO3-5813 Use new_feedback_report_url in text footer too

* AO3-5813 Make translatable app_short_name for signatures

* AO3-5813 Prefer double quotes in invite_request_declined mailer