AO3-5667 Use the reindex_world queue for the Tag:reset_filters task (#3822)
We don't want the Tag:reset_filters task to hold up normal operations
when it's being run, so we modify it to use the "reindex_world" queue
for the asynchronous FilterUpdate jobs, and use the reindexing queue
"world" instead of "background."

Because the IndexQueue doesn't normally run items added to the "world"
queue, we additionally add a rake task search:run_world_index_queue that
reads each of the world IndexQueues and runs them. This task should be
run after all of the filter updates are done and the "reindex_world"
queue is empty.