AO3-4634 All claims visible only to maintainer
If the `for_user` parameter has been removed or otherwise tampered with by a user who doesn't have permission to view all the claims, return them to the collection home page with an error message.

There is no direct link for a regular user to access all the claims in a prompt meme collection, so if we detect such a request we can reject it.