AO3-5884 Fix deleting roles when admins don't have permissions to update roles
Previously, if an admin does not have permissions to update roles, and we don't
include "roles" in the permitted attributes, the user's roles will be set to
an empty array. Now we skip setting roles entirely in that case.

Also stop skipping validations when updating users, and show validation errors
to admins, so admins can't accidentally break users by saving blank emails, etc.