AO3-5765 Remove deprecated attribute
`data-widget-id` has been deprecated for years.[1] Our timeline is properly set up, but it carries this ancillary attribute, and the deprecation warnings make it hard to wade through the output of the test suite.

There's a whole discussion about why we're making live requests to Twitter in unrelated tests, but one thing at a time.

[1] https://twittercommunity.com/t/deprecating-widget-settings/102295