AO3-6082 Use colors for RSpec and Cucumber output in GitHub Actions
By default, both RSpec and Cucumber will print colors only if the output is TTY
(as opposed to e.g. files), but the default environment of GitHub Actions does
not identify itself as running within a TTY enabled device.

So we need to use options to force color in output. For RSpec, it's easy to do
that just in CI by setting SPEC_OPTS. For Cucumber, I just do that in the
default profile, instead of adding logic to check if we're running in CI.

Also remove the deprecated "color" option of RSpec.