AO3-4008 Point kudos link in works module to kudos
Link was anchored to #comments, while #kudos is a better option.

While #kudos is not available in the template by default, current
condition for whole kudos section to be visible is for chapter to be
posted (hence work must be).
[Remark: I am not familiar enough with web development to question this
design, hence not touching it.]

On the other hand, kudos link on the works module appears only if some
kudos exist. For users to be able to give kudos, work must be posted
first.

As work can't be unposted, situation where kudos link is present but
doesn't lead anywhere shouldn't arise.

However, if this somehow does happen, that won't create a big problem as
invalid anchors are simply ignored.