AO3-4008 Scroll to comments section on click
Exchanged scrolling to feedback section with scrolling to comments

When comments are not yet loaded, next usually happens:
1) page gets scrolled to comments section - on wide screens to the
bottom of the page. Due to that user is not directly at the comments
section.
2) comments get loaded only after some time extending the page

This sequence of actions still seems like the friendliest behavior for
users, even though users will have to scroll down a bit.

Alternatives are:
1) wait until comments get loaded (AJAX) and only then scroll down - not
user friendly as navigation response is expected to be immediate.
2) scroll to comments section, wait until comments get loaded and scroll
down second time - not user friendly due to double scroll.