AO3-4008 Put commements section after kudos
This commit makes structure as following:

  #feedback
    ...
    #kudos [if visible]
    ...
    #comments
    ...
    #comments_placeholder [if visible]

This commit doesn't fully separate comments and kudos structures, as,
for example, comments and kudos errors/notices both appear before the
kudos section. They seem to be pretty tangled together. It's not a part
of this change to make something about it.

Please note that it seems currently impossible to test that navigation
links direct to correct sections on the web page.
Only possible option must be to attempt testing currently visible
objects on the screen (seems not to be possible with pure Capybara, but
should be somehow possible with javascript). Even if this is
implementable, the test, however, might be too unstable due to various
screen sizes.