AO3-6096 Handle encoded characters in Dewplayer audio URLs (#3949)
* AO3-6096 Handle encoded spaces in Dewplayer audio URLs

URL-encoded spaces in audio URLs of Dewplayer embeds get decoded when
we parse flashvars, then the URLs with unencoded spaces get rejected
as invalid by the URI class when we try to check their hosts in
MediaSanitizer. We'll instead use Addressable::URI to parse URLs,
which is more forgiving.

Also print out errored chapter IDs at the end of the rake task.

* AO3-6096 Handle empty URLs in MediaSanitizer

Previously, if the media elements has no source URLs, we rely on this
behavior of URI to return nil:

    URI("https://").host

Addressable::URI will raise an exception for the same input. So it's
better to return nil from the start.

* AO3-6096 Handle encoded and unencoded source URLs

Parse flashvars manually instead of using CGI.parse so we can be
flexible about when to unencode URLs. This is similar to how we parse
both types of URLs in EmbedSanitizer#ensure_https.

Undo all changes to MediaSanitizer.

* AO3-6096 Handle quotes in source URLs

Use Nokogiri's builder class to create the audio HTML fragment. If we
use a string fragment with single/double quotes around attribute values,
a source URL containing unescaped single/double quotes will be cut off.

Redo the changes to MediaSanitizer, this time to parse hosts out of
URLs with unescaped double quotes without errors.