AO3-4008 Divide flash_div for kudos and comments
Kudos and comments use the same div to flash their messages.

With division of comments and kudos anchors it is important to
make sure they are divided so that messages are not lost to the user.
As after removing the comment message is flashed to the user, it
should be visible at the top of the screen, not hidden from the view.

As we have two flash_divs on the same page, they need to have different
symbols attached. Current implementation of flash_div assures created
element has the same class style as provided symbol.

So to divide kudos and comments two possibilities were seen:
1) Give symbols different names - thus different class names. It
requires update of some stylesheets
2) Change flash_div method to allow custom classes

First option was chosen as it seemed more along the lines of existing
code.

Changes brought by this commit:

With enabled JS comment deletion will scroll to the deletion message,
not to the new comment field.