Cap the amount of tags you can apply to any taggable
So recently a few users have started adding massive amount of tags to things. It's causing a severe amount of issues for users. I figured I'd at least start a pull request with the idea that I would love to work on it to completion with guidance from the repository owners. I figure there's another hour of work in here at best, including:

  1. Moving to configuration file
  2. Writing data migration
  3. Dealing with the validation during update