AO3-6152 Update mimemagic
Use the latest 0.3.x version (because all the previous versions have been removed).

The gem mimemagic, that is a dependency of Rails, was using the MIT license, but distributing a file that falls under the GPL. (https://github.com/minad/mimemagic/issues/97)

The maintainer has updated the license to GPL-2.0 and yanked all the older versions of the gem.

The Archive code is already using GPL-2, so we don't need to change anything to continue using this gem, but we need to use a version that exists in rubygems so our builds don't fail.