AO3-5964 Replace extra spaces in tags with underscores
This is a one-time After task to:
- Identify tags that have extraneous spaces in their name
- Replace each space in their name with an underscore (`_`)
- Output the list of tags with their ID, old name, and new name

We have to use `update_column` because only admins are allowed to rename tags the regular way.

The way these tags have sneaked through our validation process is most likely the fact that `squish` used to use a more lax regular expression to compress blank spaces (`\s`) than the one it currently uses (`[[:space]]`).