AO3-5988 Task to fix specific noncanonical rating
We accidentally imported a large number of works with a noncanonical rating "Teen & Up Audiences".

Given that this is our culpa, we want to fix these works and give them the canonical version, "Teen And Up Audiences".

At the end of the task we make sure that no other works use the erroneous rating, and if so destroy it.