AO3-5127 Remove Mystery Series blurb leftovers
It's not precisely clear from the git log when we lost Mystery Series
blurbs, but looks like they are long gone. Series seems not to be
displayed as Mystery blurb anywhere.

Note, however, that at least title "Mystery Series" is still supposed to
be used when series are unrevealed, and there are plans for some further
improvements (see AO3-1250)