AO3-6093 Fix test step which verifies presence of input with id
As the opposite step was not run to check for presence of the same
element where it was expected in the test, it was not caught that step
implementation was faulty and test didn't perform the actual check.

To prevent similar problems in the future, test is updated and xpath
search by tag/attribute/value is generalized.

Generalized function might be converted to its own step in the future,
if more variations than 3 current existing instances are ever required.