AO3-6291 Resolve ES sorting ties
Ties resolved on Works, Bookmarks and Tags. Default Query sort is left
as is.

To resolve ties id is used in Works and Tags.
Bookmarks index contains bookmarks and bookmarkables. Bookmarks have ids
present, when bookmarkables do not and just have their _ids overridden.
Using _id is discouraged by ES, but there seems to be no other choice
for bookmarkables.