AO3-6346 HTML error pages (such as 404) for unregistered MIME types
It also prevents MissingTemplate errors that could have raised inside
of ErrorController. Previously 500 error was shown on the surface
after failing to find a non-HTML template (even when the actual error
that should be shown was 404, for example).

502 is dropped from the list of errors for ErrorController to render
because it is not found in app/views/errors/ nor in config/routes.rb.