Fix tests that write to hidden autocomplete inputs
With JS enabled, jquery.tokeninput.js hides the original inputs and
replaces them with new inputs named with the suffix "_autocomplete".
Tests should either use the right input names or use the label text.

With JS enabled, writing to an input alone does not clear the existing
selected autocomplete values. Tests need to explicitly click "delete"
on each of them. This was the case for tests that edit a work's gift
recipients.

Remove a step that checks for text inside a select ("Chapter 1").
If we want to keep it, we should use the have_select matcher.