AO3-6354 Ignore hidden elements with Capybara in feature tests (#4252)
* Update Capybara to 3.x

* Ignore hidden elements

* Fix checking textarea contents in share modals

* Fix checking CSS text in style elements

* Fix spacing in visible text to match

* Fix checking page title

* Fix test on hidden recent bookmarks

* Fix tests that wrongly matched error text in validation JS code

Error text in JS code is now considered hidden.

Previously, the invalid comment was not long enough to be rejected.

* Fix test that writes to a hidden synonym field

The autocomplete field allows just one value and is hidden after the
first selection. The old steps re-save the tag without changes so we
can remove them.

* Fix tests that write to hidden autocomplete inputs

With JS enabled, jquery.tokeninput.js hides the original inputs and
replaces them with new inputs named with the suffix "_autocomplete".
Tests should either use the right input names or use the label text.

With JS enabled, writing to an input alone does not clear the existing
selected autocomplete values. Tests need to explicitly click "delete"
on each of them. This was the case for tests that edit a work's gift
recipients.

Remove a step that checks for text inside a select ("Chapter 1").
If we want to keep it, we should use the have_select matcher.

* Remove visible: true from autocomplete finders since it's default now

* Fix comment typo

* Keep gems in Gemfile without versions

* Add Gemfile.lock changes