Use 24-hour time format by default, to improve user experience
The vast majority of countries (except USA, parts of Canada (not
even the whole country), Australia, Newzealand and the Philippines)
on the world uses the modern 24-hour format (and americans would
know it from military use in their country as well). For the vast
majority of people, the AM/PM thing is very hard to read, the
placement of 12 AM/PM especially is illogical.

Switch the default until someone makes this, or ideally the whole
date/time format, a user-customisable preference.

This solves support ticket 610765.