AO3-6460 Avoid caching stale AdminSetting
Write the modified record to the cache, instead of clearing the cache.
Then the next read will get the modified version, instead of possibly
fetching an old version from the database.

In specs, save the skin ID to the AdminSetting record instead of the
whole Skin record. A newly created Skin record will have a closed file
handle from generating a preview image and cannot be serialized for
caching (IOError: closed stream).