AO3-5901 Ensure pseud switcher shows current pseud
When we started limiting the number of pseuds in the sidebar, the way to guarantee that the current pseud would show up in the sidebar was to show it at the top of the pseud switcher.

However, [AO3-6249](https://otwarchive.atlassian.net/browse/AO3-6249) decided that the pseud switcher should always say "Pseuds" at the top.

We broke this behaviour in https://github.com/otwcode/otwarchive/pull/4554.

Fixing it in c3e4ab9c04e0299d20799f90a224b7d24b2f7a6b resulted in the current pseud not showing in the sidebar in some conditions.

Move or add the current pseud at the beginning of the list, whether or not it is present in the abbreviated list.

We are aware that this could make the abbreviated list in the pseud switcher show `ITEMS_PER_PAGE` + 1 pseuds in some cases, discuss with @sarken if you disagree.