AO3-6553 Turn homepage media list into navigation
Semantically, the list of media (plus All Fandoms) on the homepage is a
navigation element, not a menu.

The list of media used in the main site menu needs to stay a `menu`, so
we're unDRYing the partial and creating a new partial just for use on
the homepage.

The preferred element to use in modern web development is `nav`. We
don't use this anywhere else (yet), so this will be our test case.

(Per https://www.w3.org/TR/html-aria/#docconformance, `ul`s are not
allowed to have the role `navigation`.)