AO3-6552 Login link needs ARIA role (#4580)
* AO3-6552 Add menuitem ARIA role to login link

The containing paragraph has a `menu` role. This requires[1] its direct
children to have one of a few allowed roles, the most appropriate for
this case being `menuitem`.

[1]https://www.w3.org/TR/wai-aria-1.1/#menu

* AO3-6552 Add i18n for the login link

Separate commit so it can be easily removed if we don't want to i18n
partially. (I'd prefer not to i18n the entire file in this PR.)