AO3-6650 Move speculation rules next chapter prefetch (#4684)
* AO3-6650 Move speculation rules next chapter prefetch

Previously, we were adding the <script type=speculationrules> block inside the _chapter include. However, this means it was included on the "Entire Work" page, which is undesirable. (On that page, there would be no @next_chapter variable, so nothing would be rendered... and then that lack of speculation rules would be cached.)

We only want to preload the next chapter when we're looking at a single chapter in isolation. For this, it's best to put the <script type=speculationrules> in the chapters/show template.

* Add comment