AO3-6676 Remove participant memoisation
Too much of a good thing can be harmful - we don't want to memoise the participant for a collection, because the blurb is used on listings of multiple collections.

Memoising (persisting the value in a variable instead of computing it each time) is useful when we want to use the same computation during a request. Here, the computation needs to be different from one collection to the next.

The participant doesn't need to be an instance variable either. It's not what causes the bug, reusing the value is what ails us, but I have some vague memories of instance variables being less performant than plain variables, and therefore we want to prefer plain variables when it makes no difference.