AO3-6575 Turn News posts sidebar into navigation (#4591)
On the AO3 News page, the sidebar has the role `menu`, but its direct
children do not have a relevant ARIA role (e.g. `menuitem`).

The list shouldn't be a menu in the first place, but a navigation. The
preferred element to use is `nav` (rather than a `div` with the role
navigation).