AO3-6772 Remove useless html_safe call
The url is in the form of "https://archiveofourown.org/users/password/edit?reset_password_token=Token"

Nothing in this url could become mangled by getting html escaped, it's a
standard url and the Token is generated with SecureRandom.urlsafe_base64
without padding, so can only contain A-Z, a-z, 0-9, - and _.
So it does not need to be marked as html_safe.

Additionally, the admin version of this mail also doesn't call html_safe
on the url, neither does devise's own implementation of it.