AO3-6235 Allow policy_and_abuse admins to toggle invite requests in settings (#5214)
* AO3-6235 Allow policy_and_abuse admins to toggle invite requests in settings

* AO3-6235 Update SettingController rspec test

Updated an rspec test to confirm that admins with the
policy_and_abuse role can update the request_invite_enabled
AdminSetting attribute.

* AO3-6235 Refactor policy_and_abuse Admin Setting rspec test

Refactor an rspec test that validates that policy_and_abuse
admins can update permitted AdminSetting attributes.

The intent behind this refactor is to more accurately reflect
the payload the frontend is submitting to the endpoint. Before,
the test passed in boolean attributes as true/false,
ex: hide_spam: true. However, the frontend form submits these
boolean values as 0/1, and those values are saved to the db
as 0/1. The test has been updated to use 0/1 as input parameters
instead.