AO3-5542 Signup summary for challenges with no tags (#5452)
* AO3-5542 Signup summary for challenges with no tags

First, I recreated the error condition by making the gift exchange truly tagless. This also required changing the signup step for a tagless exchange to fill in both request and offer description fields, otherwise the signup was invalid.[1]

With these changes, the existing cucumber test for a tagless gift exchange started failing, so it is now a reliable test for the fix.

The fix itself required bypassing the signup summary service object entirely, and bailing out as soon as we detect that there is no topmost tag to rely on.

In that case, we don't care how many signups there are, there's no need to enqueue summary generation, since there won't be any summary.

[1] Why was it not invalid before? That's a whole other kettle of worms: we aren't successfully rejecting blank prompts in `ChallengeSignup#accepts_nested_attributes_for` as long as tags are anywhere in the picture, because of the hidden `tag_set_attributes[updated_at]` attribute (see app/views/prompts/_prompt_form_tag_options.html.erb, `tag_set_form.hidden_field :updated_at`). Monkey, circus, etc.

* AO3-5542 Linter fixes

* AO3-5542 Use heading in translation key